Implementing esp_insights_unregister_periodic_handler.
In order to properly deinitialize the ESP Insights it is necessary
to delete the periodic handler timer, stop the work queue and
deinit it too. Finally the esp_insights_entry_t structure is freed.

Please note that that it was necessary to add a modification for
the rmaker_common component (submodule) that is being reviewd at
https://github.com/espressif/esp-rainmaker-common/pull/2