esp_insights: Fixed a race condition
Timer, data_send_timer, was created after call to
esp_insights_register_periodic_handler, which queues
insights_periodic_handler to work_queue and this handler controls
this timer.
With Alexa, it was seen that post registering the handler,
work_queue thread executes the handler before the timer is created.
This led to crash.